python-weather-app/requirements.txt
# 날씨 앱 필수 패키지

# HTTP 요청을 위한 라이브러리
requests>=2.28.0

# 환경변수 로드를 위한 라이브러리 (선택사항)
python-dotenv>=0.19.0

# 테스트를 위한 패키지
pytest>=7.0.0
pytest-mock>=3.6.0

# 코드 포맷팅 (개발용)
black>=22.0.0
flake8>=4.0.0